% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{adjustProfiles,Simulator-method}
\alias{adjustProfiles,Simulator-method}
\title{adjustProfiles}
\usage{
\S4method{adjustProfiles}{Simulator}(object, simulation, profiles, step)
}
\arguments{
\item{object}{Instance of class \linkS4class{Simulator}.}

\item{simulation}{Instance of class \linkS4class{Simulation}.}

\item{profiles}{Data frame containing the profile type associated to each
ID.}

\item{step}{It accepts two options \emph{Effect} and \emph{Groups} depending
on the part of the process where this method is called. Check 'Simulation.R'.}
}
\value{
By default it returs the profiles data frame without modifications.
}
\description{
For internal use. Allows every omic class to adjust the profiles matrix
(simulation settings) to use, according to its needs.
}
