% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{adjustProfiles}
\alias{adjustProfiles}
\title{adjustProfiles}
\usage{
adjustProfiles(object, simulation, profiles, step)
}
\arguments{
\item{object}{Instance of simulator class.}

\item{simulation}{Instance of simulation class.}

\item{profiles}{Data.frame containing the generated profile configuration for the
simulator.}

\item{step}{Two possible options "Effect" and "Groups" indicating in which point
of the settings simulation is called.}
}
\value{
The data.frame profile with the necessary modifications depending on
each child class.
}
\description{
Modify the configuration profile generated when initializing simulation settings
to allow for certain constrainsts if necessary.
}
