% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{associationList}
\alias{associationList}
\title{Data to showcase scRNA and scATAC-seq association}
\format{
A dataframe with two columns and rows according to gene/feature relationships
 \describe{
     \item{Peak_ID}{ATAC chromosomic positions associated to genes}
     \item{Gene_ID}{RNA genes associated to peaks}
     }
     
 @source {Created in-house to serve as an example}
 @examples
}
\usage{
data(associationList)
}
\description{
Data to showcase scRNA and scATAC-seq association
}
\keyword{datasets}
