% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_coexpression.R
\name{calculate_mean_per_list_df}
\alias{calculate_mean_per_list_df}
\title{calculate_mean_per_list_df}
\usage{
calculate_mean_per_list_df(df, named_lists)
}
\arguments{
\item{df}{dataframe of expression where columns are cells}

\item{named_lists}{list of which cells belong to each celltype}
}
\description{
Helper function to calculate mean expression per celltype
}
\examples{
rna <- data.frame(c1 = c(1.5, 15.5, 3.5, 20.5), c2 = c(2, 15, 4, 20), 
          c3 = c(10, 1, 12, 13), c4 = c(11, 1, 13, 14))
cell_types <- list("ct1" = c(1,2), "ct2" = c(3, 4))
calculate_mean_per_list_df(rna, cell_types)
}
