% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_coexpression.R
\name{check_patterns}
\alias{check_patterns}
\title{check_patterns}
\usage{
check_patterns(patterns_ret)
}
\arguments{
\item{patterns_ret}{tibble of TRUE FALSE values}
}
\value{
list of indices where the rows are opposite
}
\description{
Function to check if the TRUE FALSE patterns have at least two rows that are
opposite, we need this to be able to generate repressor regulators
}
\examples{
patterns <- tibble(one = c(TRUE, FALSE, TRUE, FALSE), 
              two = c(TRUE, TRUE, TRUE, TRUE), 
              three = c(FALSE, TRUE, FALSE, TRUE), 
              four = c(FALSE, TRUE, TRUE, TRUE))
opposite_indices <- check_patterns(patterns)

}
