% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{CollinearityFilter}
\alias{CollinearityFilter}
\title{Title}
\usage{
CollinearityFilter(data, reg.table, correlation = 0.8, action = c("random",
  "mean"))
}
\arguments{
\item{action}{}
}
\description{
Title
}
