% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{CorreAction}
\alias{CorreAction}
\title{Title}
\usage{
CorreAction(correlacionados, data, reg.table, omic, action = c("mean",
  "random"), nombre = "mc1")
}
\arguments{
\item{nombre}{}
}
\description{
Title
}
