% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{discretize}
\alias{discretize}
\title{Discretize ChIP-Seq counts to simulate a binary dataset}
\usage{
discretize(df)
}
\arguments{
\item{simulated}{A MOSimulated object}
}
\value{
A ChIP-seq dataframe of 0 and 1
}
\description{
Discretize ChIP-Seq counts to simulate a binary dataset
}
\examples{
discrete_ChIP <- discretize(rnaseq_simulated)

}
