% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{experimentalDesign}
\alias{experimentalDesign}
\title{Retrieves the experimental design}
\usage{
experimentalDesign(simulation)
}
\arguments{
\item{simulation}{A MOSimulation object}
}
\value{
A data frame containing the experimental design used to simulate the
  data.
}
\description{
Retrieves the experimental design
}
\examples{

omic_list <- c("RNA-seq")
rnaseq_simulation <- mosim(omics = omic_list)
# This will be a data frame with RNA-seq counts

design_matrix <- experimentalDesign(rnaseq_simulation)

}
