% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{GetGLM}
\alias{GetGLM}
\title{Genome-wide Generalized Linear Models}
\usage{
GetGLM(GeneExpression, associations, data.omics, edesign, cont.var = "Time",
  degree = 1, Res.df = 5, alfa = 0.05, MT.adjust = "none",
  family = negative.binomial(theta = 10), stepwise = "backward",
  interactions.exp = TRUE, interactions.reg = TRUE, min.variation = 0,
  correlation = 0.9, action = "mean", min.obs = 10, epsilon = 1e-05,
  mc.cores = 1)
}
\arguments{
\item{GeneExpression}{Data frame containing gene expression data with genes in rows and
experimental samples in columns. Row names must be the gene IDs.}

\item{associations}{List where each element correspond to a different omic data type (miRNAs,
transcription factors, methylation, etc.). The names of the list will be the omics. Each element
is a data frame with 2 columns (optionally 3) describing the potential interactions between genes
and regulators for that omic. First column must contain the genes (or features in
GeneExpression object), second column must contain the regulators, and an additional column can
be added to describe the type of interaction (e.g., for methylation, if a CpG site is located in
the promoter region of the gene, in the first exon, etc.).}

\item{data.omics}{List where each element correspond to a different omic data type (miRNAs,
transcription factors, methylation, etc.). The names of the list will be the omics. Each element
is a data matrix with omic regulators in rows and samples in columns.}

\item{edesign}{Data frame describing the experimental design. Rows must be the samples (columns
in GeneExpression) and columns must be the experimental variables to be included in the model
(e.g. time, treatment, etc.).}

\item{cont.var}{Name of the column in edesign that is to be considered as a continuous
explanatory variable in the regression model. NULL if edesign does not contain any continuous
variables. By default, "Time".}

\item{degree}{If cont.var is not NULL, non-linear (polynomial) relationships between the cont.var
and the gene expression must be studied. By default, degree = 1 and must be an integer.
A higher number will allow for in quadratic, cubic, etc. terms for cont.var in the model.}

\item{Res.df}{Number of degrees of freedom in the residuals. By default, 5. Increasing
Res.df will increase the power of the statistical model.}

\item{alfa}{Significance level. By default, 0.05.}

\item{MT.adjust}{Multiple testing correction method to be used within the stepwise variable
selection procedure. By default, "none". See the different options in ?\code{p.adjust}.}

\item{family}{Error distribution and link function to be used in the model (see ?\code{glm}
for more information). By default, negative.binomial(theta=10).}

\item{stepwise}{Stepwise variable selection method to be applied. It can be one of: "backward"
(default), "forward", "two.ways.backward" or "two.ways.forward".}

\item{interactions.exp}{If TRUE (default), interactions among the experimental variables
are included in the model.}

\item{interactions.reg}{If TRUE (default), interactions between regulators and experimental
variables are included in the model.}

\item{min.variation}{### if NULL, it computes "sd"}

\item{min.obs}{}
}
\description{
\code{GetGLM} fits a regression model for all the genes in the data set to identify
the experimental variables and potential regulators that show a significant effect on
the expression of each gene.
}
