% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{IDfromGenes,Simulator-method}
\alias{IDfromGenes,Simulator-method}
\title{IDfromGenes}
\usage{
\S4method{IDfromGenes}{Simulator}(object, geneNames, simplify = TRUE)
}
\arguments{
\item{object}{Instance of class \linkS4class{Simulator}.}

\item{geneNames}{Character vector of the genes to look for in the association table.}

\item{simplify}{Return only the genes IDs or a table containing both types of identifiers.}
}
\value{
Depending on \emph{simplify} parameter it would be a character vector or a
data frame containing columns \emph{ID} and \emph{Gene}.
}
\description{
Returns the regulator IDs associated to a particular set of genes identifiers.
}
