% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{IDfromGenes}
\alias{IDfromGenes}
\title{IDfromGenes}
\usage{
IDfromGenes(object, geneNames, simplify = TRUE)
}
\arguments{
\item{object}{Instance of a Simulator class}

\item{geneNames}{Names of genes to look up in the association table.}
}
\value{
IDs corresponding to the genes.
}
\description{
Method for transforming genes to IDs (regions, miRNA...)
}
