% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{IDtoGenes,Simulator-method}
\alias{IDtoGenes,Simulator-method}
\title{IDtoGenes}
\usage{
\S4method{IDtoGenes}{Simulator}(object, idNames, simplify = TRUE)
}
\arguments{
\item{object}{Instance of class \linkS4class{Simulator}.}

\item{idNames}{Character vector of the regulator IDs to look for in the association table.}

\item{simplify}{Return only the regulator IDs or a table containing both types of identifiers.}
}
\value{
Depending on \emph{simplify} parameter it would be a character vector or a
data frame containing columns \emph{ID} and \emph{Gene}.
}
\description{
Returns the gene identifiers associated to a particular set of regulator IDs.
}
