% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{IDtoGenes}
\alias{IDtoGenes}
\title{IDtoGenes}
\usage{
IDtoGenes(object, idNames, simplify = TRUE)
}
\arguments{
\item{object}{Instance of a simulator class.}

\item{idNames}{IDs to look up in the association table.}
}
\value{
Genes corresponding to the IDs.
}
\description{
Method for transforming IDs (regions, miRNAs, ...) to genes
}
