% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Simulator.R, R/Simulation.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{initialize,Simulator-method}
\alias{initialize,Simulator}
\alias{initialize,Simulation-method}
\alias{initialize,Simulation}
\title{Simulate [internal use]}
\usage{
simulate(object, ...)

\S4method{initialize}{Simulator}(.Object, ...)

\S4method{initialize}{Simulation}(.Object, ...)
}
\arguments{
\item{object}{Object of class \linkS4class{Simulator} or \linkS4class{Simulation}.}

\item{...}{Extra parameters for extensibility.}

\item{simulation}{Only in \linkS4class{Simulator} class. Initialized instance of Simulation class}
}
\value{
A \linkS4class{Simulator} object containing the simulation data inside @simData
 if called on a \linkS4class{Simulator} class, or a \linkS4class{Simulation} object with all @simulators
 containing simulated data.
}
\description{
This generic can dispatch methods for two types of classes:
\describe{
     \item{\linkS4class{Simulation}}{in which case triggers the simulation on every simulator
 loaded.}
     \item{\linkS4class{Simulator}}{performs the simulation based on already initialized simulation
 settings.}
}
}
