% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{initializeData,Simulator-method}
\alias{initializeData,Simulator-method}
\title{initializeData}
\usage{
\S4method{initializeData}{Simulator}(object, simulation)
}
\arguments{
\item{object}{Instance of \linkS4class{Simulator} class.}

\item{simulation}{Initialized instance of \linkS4class{Simulation} class}
}
\value{
An object of class \linkS4class{Simulator} with the @data slot filled
with the correct initial data structure and values.
}
\description{
Generates the initial sample of data. It copies the initial data for every
group of the experimental design, and ensures to change the count values
for those genes (and associated regulators with an active effect on it)
that are marked as DE with only flat profiles in all groups.
}
