% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{is.declared}
\alias{is.declared}
\title{Check if a variable is declared.}
\usage{
is.declared(object, key = NULL)
}
\arguments{
\item{object}{Variable name to check}

\item{key}{Optional key to check inside object.}
}
\value{
TRUE or FALSE indicating if the variable is initialized & non-empty.
}
\description{
Check if a variable is declared.
}
