% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{locGRanges}
\alias{locGRanges}
\title{locGRanges}
\usage{
locGRanges(object, locs)
}
\arguments{
\item{object}{Instance of a SimulatorRegion class}

\item{locs}{Positions of locations}
}
\value{
GRanges object with 1 as seqname and locs as starting positions.
}
\description{
Creates a very basic GRanges object based on locations
}
