% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_coexpression.R
\name{make_association_dataframe}
\alias{make_association_dataframe}
\title{make_association_dataframe}
\usage{
make_association_dataframe(
  omics,
  associationList = NULL,
  minFC = 0.25,
  maxFC = 4,
  group = 1,
  genereggroup
)
}
\arguments{
\item{omics}{}

\item{associationList}{OPTIONAL. A 2 columns dataframe reporting peak ids 
related to gene names. If not provided, no table of association between
scRNA and scATAC is outputted.}

\item{minFC}{OPTIONAL. Threshold of FC below which are downregulated, by 
default 0.25}

\item{maxFC}{OPTIONAL. Threshold of FC abofe which are upregulated, by default 4}

\item{group}{Group from which we are generating the association dataframe}

\item{genereggroup}{list of elements to generate the association dataframe
such as clusters of each omic, indices of opposite clusters, which
genes are activated, repressed, behavior of the features etc.}
}
\value{
a dataframe with all the information the user needs about each gene
}
\description{
make_association_dataframe
}
