% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\docType{package}
\name{MOSim-package}
\alias{MOSim}
\alias{MOSim-package}
\title{MOSim}
\description{
Multiomics simulation package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ConesaLab/MOSim}
  \item Report bugs at \url{https://github.com/ConesaLab/MOSim/issues}
}

}
\author{
\strong{Maintainer}: Sonia Tarazona \email{sotacam@gmail.com}

Authors:
\itemize{
  \item Carolina Monzó \email{carolmonzoc@gmail.com}
  \item Carlos Martínez \email{cmarmir@gmail.com}
}

}
\keyword{internal}
