% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{omicSettings}
\alias{omicSettings}
\title{Retrieves the settings used in a simulation}
\usage{
omicSettings(simulation, omics = NULL, association = FALSE,
  reverse = FALSE, only.linked = FALSE, prefix = FALSE)
}
\arguments{
\item{simulation}{A Simulation object.}

\item{omics}{List of omics to retrieve the settings.}

\item{association}{A boolean indicating if the association must also be returned for the regulators.}

\item{reverse}{A boolean, swap the column order in the association list in case we want to use the
output directly and the program requires a different ordering.}

\item{only.linked}{Return only the interactions that have an effect.}
}
\value{
A list containing a data frame with the settings used to simulate each of the indicated omics.
If association is TRUE, it will be a list with 2 keys: 'associations' and 'settings', with each one being
a list containing the information for the selected omics.
}
\description{
Retrieves the settings used in a simulation
}
