% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_coexpression.R
\name{order_FC_forMatrix}
\alias{order_FC_forMatrix}
\title{order_FC_forMatrix}
\usage{
order_FC_forMatrix(A, B, C, D)
}
\arguments{
\item{A}{Vector of c("Up", "Down", "NE) from the Gene or Peak_DE extracted
from the association matrix}

\item{B}{Calculated vector of Up FC values}

\item{C}{Calculated vector of Down FC values}

\item{D}{Calculated vector of NE FC values}
}
\description{
Function to sort the FC values according to the genes that must be up or 
downregulated
}
\examples{
DE <- c("Up", "Up", "Up", "Down", "Down", "NE", "NE", "NE", "NE", NA, NA, NA)
Up_FCvec <- c(1, 1, 1)
Down_FCvec <- c(2, 2)
notDE_FCvec <- c(2, 2, 2, 2)
FC_vec <- order_FC_forMatrix(DE, Up_FCvec, Down_FCvec, notDE_FCvec)

}
