% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{param_estimation}
\alias{param_estimation}
\title{param_estimation}
\usage{
param_estimation(omics, cellTypes, numberCells = NULL, mean = NULL, sd = NULL)
}
\arguments{
\item{omics}{named list containing the omics to simulate as names, which can 
be "scRNA-seq" or "scATAC-seq, and the input count matrix}

\item{cellTypes}{list where the i-th element of the list contains the column 
indices for i-th cell type. List must be a named list.}
}
\value{
a named list with simulation parameters for each omics as values
}
\description{
Evaluate the users parameters for single cell simulation and use SPARSim
to simulate the main dataset
}
