% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{plotGeneRegu}
\alias{plotGeneRegu}
\title{Title}
\usage{
plotGeneRegu(x.points, geneValues, reguValues, geneErrorValues, reguErrorValues,
  col = c(1, 2), xlab = "", yylab = c("right", "left"), pch = c(16, 17),
  main = "", numLines = 0, x.names = NULL, yleftlim, yrightlim,
  condi1 = NULL, condi2)
}
\arguments{
\item{yrightlim}{}
}
\description{
Title
}
