% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MORE_GLM.R
\name{plotGLM}
\alias{plotGLM}
\title{Title}
\usage{
plotGLM(GLMoutput, gene, regulator = NULL, reguValues = NULL,
  plotPerOmic = TRUE, fittedModel = FALSE, replicates = FALSE,
  gene.col = 4, regu.col = NULL, xlab = "", cont.var = "Time",
  cond2plot = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
