% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{postSimulation,Simulator-method}
\alias{postSimulation,Simulator-method}
\title{postSimulation}
\usage{
\S4method{postSimulation}{Simulator}(object, simulation)
}
\arguments{
\item{object}{Instance of class \linkS4class{Simulator}.}

\item{simulation}{Instance of class \linkS4class{Simulation}.}
}
\value{
Object of class \linkS4class{Simulation} with the simulated data (@simData)
correctly formatted.
}
\description{
Method to make final modifications required by the different omics. Like rounding
the final count values or renaming columns for making them more human-readable.
}
\details{
For internal use only.
}
