% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{postSimulation}
\alias{postSimulation}
\title{postSimulation}
\usage{
postSimulation(object, simulation)
}
\arguments{
\item{object}{Instance of Simulator class.}

\item{simulation}{Instance of Simulation class.}
}
\value{
An instance of Simulator class with modified options.
}
\description{
Method called after performing the simulation, allowing different actions
like adjusting depth, rounding values and so on, on a simulator basis.
}
