% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGLM_function.R
\name{PreviousBackward}
\alias{PreviousBackward}
\title{Title}
\usage{
PreviousBackward(y, d, alfa, family, epsilon, stepwise, gdl.max,
  MT.adjust = "fdr", iter.max = 100)
}
\arguments{
\item{iter.max}{}
}
\description{
Title
}
