% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{regionNames}
\alias{regionNames}
\title{regionNames}
\usage{
regionNames(object, chrNumber, start, end = NULL)
}
\arguments{
\item{object}{Instace of a SimulatorRegion class}

\item{chrNumber}{IDs of the chromosomes}

\item{start}{Starts positions of the regions}

\item{end}{End positions of the regions}
}
\value{
A character vector with the format "<chr>_<start>_<end>"
}
\description{
Creates a valid string of a region
}
