% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\docType{data}
\name{sampleData}
\alias{sampleData}
\title{Default data}
\format{
An object of class \code{list} of length 6.
}
\usage{
sampleData
}
\description{
Dataset with base counts and id-gene tables.
}
\details{
List with 6 elements:
\describe{
  \item{SimRNAseq}{\describe{
  \item{data}{Dataframe with base counts with gene id as rownames.}
  \item{geneLength}{Length of every gene.}}}
  \item{SimChIPseq}{\describe{
  \item{data}{Dataframe with base counts with regions as rownames.}
  \item{idToGene}{Dataframe with region as "ID" column and gene name on "Gene" column.}}}
  \item{SimDNaseseq}{\describe{
  \item{data}{Dataframe with base counts with regions as rownames.}
  \item{idToGene}{Dataframe with region as "ID" column and gene name on "Gene" column.}}}
  \item{SimMiRNAseq}{\describe{
  \item{data}{Dataframe with base counts with miRNA id as rownames.}
  \item{idToGene}{Dataframe with miRNA as "ID" column and gene name on "Gene" column.}}}
  \item{SimMethylseq}{\describe{
  \item{idToGene}{Dataframe with region as "ID" column and gene name on "Gene" column.}}}
  \item{CpGisland}{Dataframe of CpG to be used as initialization data, located on "Region" column}
}
}
\keyword{datasets}
