% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_omicData}
\alias{sc_omicData}
\title{sc_omicData}
\usage{
sc_omicData(omics, data = NULL)
}
\arguments{
\item{data}{A user input matrix with genes (peaks in case of scATAC-seq) 
as rows and cells as columns. Alternatively MOSim allows user to estimate 
the input parameters from an existing count table by typing 'example_matrix'}

\item{omic}{A string which can be either "scRNA-seq" or "scATAC-seq"}
}
\value{
a named list with omics type as name and the count matrix as value
}
\description{
Checks if the user defined data is in the correct format, or loads
the pbmc dataset from https://satijalab.org/seurat/articles/pbmc3k_tutorial.html
}
\examples{

scRNAseq <- sc_omicData("scRNA-seq")
scATACseq <- sc_omicData("scATAC-seq")
scRNAseq_user <- sc_omicData("scRNA-seq", count_matrix) 
}
