% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_omicResults}
\alias{sc_omicResults}
\title{sc_omicResults}
\usage{
sc_omicResults(sim)
}
\arguments{
\item{sim}{a simulated object from sc_mosim function}
}
\value{
list of seurat objects with simulated data
}
\description{
sc_omicResults
}
\examples{

cell_types <- list('Treg' = c(1:10),'cDC' = c(11:20),'CD4_TEM' = c(21:30),
'Memory_B' = c(31:40))
omicsList <- sc_omicData(list("scRNA-seq"))
sim <- sc_mosim(omicsList, cell_types)
res <- sc_omicResults(sim)
}
