% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_omicSettings}
\alias{sc_omicSettings}
\title{sc_omicSettings}
\usage{
sc_omicSettings(sim, TF = FALSE)
}
\arguments{
\item{sim}{a simulated object from sc_mosim function}

\item{TF}{OPTIONAL default is FALSE, if true, extract TF association matrix}
}
\value{
list of Association matrices explaining the effects of each
regulator to each gene
}
\description{
sc_omicSettings
}
\examples{

cell_types <- list('Treg' = c(1:10),'cDC' = c(11:20),'CD4_TEM' = c(21:30),
'Memory_B' = c(31:40))
omicsList <- sc_omicData(list("scRNA-seq"))
sim <- sc_mosim(omicsList, cell_types)
res <- sc_omicSettings(sim)
}
