% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_omicSim}
\alias{sc_omicSim}
\title{sc_omicSim}
\usage{
sc_omicSim(
  sim,
  cellTypes,
  totalFeatures = NULL,
  regulatoreEffect = NULL,
  associationList = NULL
)
}
\arguments{
\item{sim}{named list containing the omic simulated as names ("scRNA-seq" and 
"scATAC-seq") , and seurat objects as values.}

\item{cellTypes}{list where the i-th element of the list contains the column 
indices for i-th experimental conditions. List must be a named list.}

\item{totalFeatures}{OPTIONAL. Numeric value. Total number of features for 
the regulatory omic.If not provided it uses the same amount of features 
    used for the simulation.}

\item{associationList}{OPTIONAL. A 2 columns dataframe reporting peak ids and 
gene names. If not provided the code uses our own associationlist derived 
    from hg19.}

\item{regulatorEffect}{OPTIONAL. Named list of length 3 where the user can 
pass the percentage of activators, repressors and NE he wants as output. 
    If not provided the function outputs the dataframe without sub-setting 
        it according to percentages.}
}
\value{
named list containing a 3 columns dataframe (peak_id, activity, 
    cell_type), one per each couple of \code{cellTypes}.
}
\description{
Defines regulatory functions of features for regulatory omics
}
\examples{


omic_list <- sc_omicData(c("scRNA-seq","scATAC-seq"))
cell_types <- list(cellA = c(1:20), cellB = c(161:191))
sim <-scMOSim(omic_list, cell_types, numberCells = c(10,20), mean = c(2*10^6, 2*10^3), sd = c(10^3, 10^2))
cell_types <- list(cellA= c(1:10), cellB = c(11:30))
regulatorEffect = list('activator' = 0.8,'repressor' = 0.1,'NE' = 0.1)
sc_omicSim(sim, cell_types, totalFeatures = 500, regulatoreEffect = regulatorEffect)

}
