% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_param_groups}
\alias{sc_param_groups}
\title{sc_param_groups}
\usage{
sc_param_groups(
  omics,
  cellTypes,
  param_est_list,
  diffGenes = c(0.2, 0.2),
  minFC = 0.25,
  maxFC = 4
)
}
\arguments{
\item{omics}{named list containing the omics to simulate as names, which can 
be "scRNA-seq" or "scATAC-seq".}

\item{cellTypes}{list where the i-th element of the list contains the column 
indices for i-th cell type. List must be a named list.}

\item{param_est_list}{Simulation parameters object from first experimental group}

\item{diffGenes}{If number groups > 1, Percentage DE genes to simulate. 
Can be a vector with absolute genes for Up and Down ex: c(250, 500) or a 
percentage for up, down by default: c(0.2, 0.2). The rest will be NE}

\item{minFC}{Threshold of FC below which are downregulated, by default 0.25}

\item{maxFC}{Threshold of FC above which are upregulated, by default 4}
}
\value{
a named list with simulation parameters for each omics as values.
}
\description{
Simulate parameters for simulation of diferentially expressed groups. Get
parameters from first experimental group, create a vector of percentage
of up and down regulated, create parameters for new DE genes and simulate
}
\examples{
omicsList <- sc_omicData(list("scRNA-seq"))
cell_types <- list(cellA = c(1:20), cellB = c(161:191))
estimated_params <- sc_param_estimation(omicsList, cell_types)
estimated_groups <- sc_param_groups(omics_list, cell_types, estimated_params,
                        diffGenes = c(0.2, 0.2), minFC = 0.25, maxFC = 4)

}
