% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scatac}
\alias{scatac}
\title{Data to test scMOSim}
\format{
A seurat Object, subset from seuratData with ATAC
 \describe{
     \item{assays}{ATAC expression values}
     \item{meta.data}{annotations of celltypes}
     }
     
 @source {https://github.com/satijalab/seurat-data, we took 11 cells 
 from each of 4 celltypes}
}
\usage{
data("scatac")
}
\description{
Data to test scMOSim
}
