% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{scOmicResults}
\alias{scOmicResults}
\title{scOmicResults}
\usage{
scOmicResults(sim, TF = FALSE, TFlist = NULL)
}
\arguments{
\item{sim}{a simulated object from scMOSim function}

\item{TF}{(optional) default is FALSE, if true, extract TF dataframe}

\item{TFlist}{(optional), default is NULL. If a vector of gene names
is given, they are extracted by the results object. If no vector
is given, using the list of human TF from 
{http://humantfs.ccbr.utoronto.ca/download.php}}
}
\value{
list of seurat objects with simulated data
}
\description{
scOmicResults
}
\examples{

cell_types <- list('Treg' = c(1:10),'cDC' = c(11:20),'CD4_TEM' = c(21:30),
'Memory_B' = c(31:40))
omicsList <- sc_omicData(list("scRNA-seq"))
sim <- scMOSim(omicsList, cell_types)
res <- scOmicResults(sim)
}
