% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{scOmicSettings}
\alias{scOmicSettings}
\title{scOmicSettings}
\usage{
scOmicSettings(sim)
}
\arguments{
\item{sim}{a simulated object from scMOSim function}
}
\value{
list of Association matrices explaining the effects of each
regulator to each gene
}
\description{
scOmicSettings
}
