% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{simDebug,Simulator-method}
\alias{simDebug,Simulator-method}
\title{simDebug}
\usage{
\S4method{simDebug}{Simulator}(object, simulation, method, ...)
}
\arguments{
\item{object}{Instance of class \linkS4class{Simulator}.}

\item{simulation}{Instance of class \linkS4class{Simulation}.}

\item{method}{Name of the method calling the function.}

\item{...}{Values to keep track of.}
}
\description{
Helper function to trace how the different values are modified during
the execution of the algorithm. It is only intended for testing purposes
and as a final user it should not be used.
}
