% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimMethylseq-class}
\alias{SimMethylseq-class}
\title{Class to simulate Methyl-seq data.}
\description{
Class to simulate Methyl-seq data.
}
\section{Slots}{

\describe{
\item{\code{nCpG}}{numeric. Number of CpG sites to simulate.}

\item{\code{pSuccessMethReg}}{numeric. Probability of success in methylated region.}

\item{\code{pSuccessDemethReg}}{numeric. Probability of success in non methylated region}

\item{\code{errorMethReg}}{numeric. Error rate in methylated region}

\item{\code{errorDemethReg}}{numeric. Error rate in methylated region}

\item{\code{nReadsMethReg}}{numeric. Mean number of reads in methylated region.}

\item{\code{nReadsDemethReg}}{numeric. Mean number of reads in non methylated regions.}

\item{\code{phaseDiff}}{numeric. Phase difference in the differentially methylated regions between two samples}

\item{\code{balanceHypoHyper}}{numeric. Balance of hypo/hyper methylation}

\item{\code{ratesHMMMatrix}}{numeric. Matrix of values that describes the exponential
decay functions that define the distances between CpG values.}

\item{\code{distType}}{character. Distribution used to generate replicates:}

\item{\code{transitionSize}}{numeric.}

\item{\code{PhiMeth}}{matrix. Transition matrix for CpG locations.}

\item{\code{PhiDemeth}}{matrix. <Not used>}

\item{\code{typesLocation}}{numeric. <Not used>}

\item{\code{returnValue}}{character. Selected column:}

\item{\code{betaThreshold}}{numeric. Beta threshold value used to calculate M values.}
}}

