% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Simulator.R, R/Simulation.R
\docType{methods}
\name{simSettings}
\alias{simSettings}
\alias{simulate,Simulator-method}
\alias{simulate,Simulation}
\alias{simulate,Simulation-method}
\title{Simulate [internal use]}
\usage{
simSettings(object)

\S4method{simulate}{Simulator}(object, simulation)

\S4method{simulate}{Simulation}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{Simulator} or \linkS4class{Simulation}.}
}
\description{
This generic can dispatch methods from two types of classes:
\describe{
     \item{\linkS4class{Simulation}}{show the general simulation settings (experimental design, default depth...)}
     \item{\linkS4class{Simulator}}{individual omic options (depth, regulator effect...)}
}
}
