% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\docType{methods}
\name{simulateParams,Simulator-method}
\alias{simulateParams,Simulator-method}
\title{simulateParams}
\usage{
\S4method{simulateParams}{Simulator}(object, simulation, counts, profiles, ids)
}
\arguments{
\item{object}{Instance of class \linkS4class{Simulator}.}

\item{simulation}{Instance of class \linkS4class{Simulation}.}

\item{counts}{Counts taken from the initial data.}

\item{profiles}{Types of profiles to simulate.}

\item{ids}{IDs associated. Needed in some simulators.}
}
\value{
A list containing the following items:
\describe{
     \item{randomCounts}{numeric vector containing random count values.}
     \item{noiseValues}{numeric vector containing noise values generated with the noise function and parameters specified.}
     \item{m}{numeric vector of lower values comparing original and random counts.}
     \item{M}{numeric vector of maximum values comparing original and random counts.}
}
}
\description{
For internal use. Creates the values to be replaced in the formulas used
to simulate the profile values, including noise.
}
