% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{simulateParams}
\alias{simulateParams}
\title{simulateParams}
\usage{
simulateParams(object, simulation, counts, profiles, group, ids, ...)
}
\arguments{
\item{object}{Instance of a simulator class.}

\item{simulation}{Instance of a simulation class}

\item{counts}{Initial count values to simulate.}

\item{profiles}{Generated simulation settings for the group being simulated.}

\item{group}{Group being simulated.}

\item{ids}{ID of every row.}
}
\value{
List with the following elements:
\describe{
 \item{}{}
}

TODO: MODIFY THE GENERIC
}
\description{
Method to allow individual simulator classes to change the default behaviour
of generating noise and time coefficients.
}
