% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SimulatorRegion-class}
\alias{SimulatorRegion-class}
\title{Virtual class containing general methods for simulators based on
regions of the chromosomes, like DNase-seq, ChIP-seq or Methyl-seq}
\description{
Virtual class containing general methods for simulators based on
regions of the chromosomes, like DNase-seq, ChIP-seq or Methyl-seq
}
\section{Slots}{

\describe{
\item{\code{locs}}{Vector containing the list of locations of the sites.}

\item{\code{locsName}}{Type of the site to simulate, only for debug.}

\item{\code{splitChar}}{Character symbol used to split identifiers in chr/start/end}
}}

