% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsim_functions.R
\name{sparsim_estimate_parameter_from_data}
\alias{sparsim_estimate_parameter_from_data}
\title{Estimate SPARSim simulation parameter from a given count table}
\usage{
sparsim_estimate_parameter_from_data(raw_data, norm_data, conditions)
}
\arguments{
\item{raw_data}{count matrix (gene on rows, samples on columns) containing raw count data}

\item{norm_data}{count matrix (gene on rows, samples on columns) containing normalized count data}

\item{conditions}{list where the i-th element of the list contains the column indices for i-th experimental conditions. List must be a named list.}
}
\value{
A SPARSim simulation parameters
}
\description{
Function to estimate SPARSim simulation parameters (intensity, variability and library sizes) from a real count table.
If the real count table contains more than one experimental condition, it is possible to estimate the parameters for each experimental condition.
}
\details{
This function is a copy of the `SPARSIM_estimate_parameter_from_data` function from the
`SPARSim` package (v0.9.5), originally developed by Giacomo Baruzzo,
Ilaria Patuzzi, Barbara Di Camillo (2020). The original package is licensed
under the GPL-3 license.
}
