% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGLM_function.R
\name{stepback.gdl}
\alias{stepback.gdl}
\title{Title}
\usage{
stepback.gdl(y, d.ini, d.res, alfa, family, epsilon, gdl.max, tmax,
  MT.adjust = "fdr")
}
\arguments{
\item{MT.adjust}{}
}
\description{
Title
}
