% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGLM_function.R
\name{StepwiseProcedure}
\alias{StepwiseProcedure}
\title{Title}
\usage{
StepwiseProcedure(stepwise, y, x, family, alfa, epsilon, MT.adjust = "fdr",
  iter.max = 100)
}
\arguments{
\item{iter.max}{}
}
\description{
Title
}
