% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TF_human}
\alias{TF_human}
\title{Data to extract human TF}
\format{
vector of gene names
 \describe{
     \item{vector}{gene names corresponding to TF}
     }
 @source {http://humantfs.ccbr.utoronto.ca/download.php}
}
\usage{
data("TF_human")
}
\description{
Data to extract human TF
}
