% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGLM_function.R
\name{two.ways.stepback.gdl}
\alias{two.ways.stepback.gdl}
\title{Title}
\usage{
two.ways.stepback.gdl(y, d.ini, d.res, alfa, family, epsilon, gdl.max,
  MT.adjust = "fdr", iter.max = 100)
}
\arguments{
\item{iter.max}{}
}
\description{
Title
}
