% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeGLM_function.R
\name{two.ways.stepbackMOD}
\alias{two.ways.stepbackMOD}
\title{Title}
\usage{
two.ways.stepbackMOD(y, d, alfa, family, epsilon, MT.adjust = "fdr",
  iter.max = 100)
}
\arguments{
\item{iter.max}{}
}
\description{
Title
}
