% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{experimentalDesign}
\alias{experimentalDesign}
\title{Retrieves the experimental design}
\usage{
experimentalDesign(simulation)
}
\arguments{
\item{simulation}{}
}
\value{
A Simulation object.
}
\description{
Retrieves the experimental design
}
