% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Simulator.R
\docType{methods}
\name{simulate}
\alias{simulate}
\alias{simSettings}
\alias{initializeData}
\alias{postSimulation}
\alias{IDfromGenes}
\alias{IDtoGenes}
\alias{simulateParams}
\alias{adjustProfiles}
\alias{locGRanges}
\alias{regionNames}
\alias{initialize,MOSimulator-method}
\alias{initializeData,MOSimulator-method}
\alias{simulate,MOSimulator-method}
\alias{postSimulation,MOSimulator-method}
\alias{IDfromGenes,MOSimulator-method}
\alias{IDtoGenes,MOSimulator-method}
\alias{simulateParams,MOSimulator-method}
\alias{adjustProfiles,MOSimulator-method}
\alias{show,MOSimulator-method}
\alias{simSettings,MOSimulator-method}
\title{Simulate [internal use]}
\usage{
simulate(object, ...)

simSettings(object)

initializeData(object, simulation)

postSimulation(object, simulation)

IDfromGenes(object, geneNames, simplify = TRUE)

IDtoGenes(object, idNames, simplify = TRUE)

simulateParams(object, simulation, counts, profiles, group, ids, ...)

adjustProfiles(object, simulation, profiles, step)

locGRanges(object, locs)

regionNames(object, chrNumber, start, end = NULL)

\S4method{initialize}{MOSimulator}(.Object, ...)

\S4method{initializeData}{MOSimulator}(object, simulation)

\S4method{simulate}{MOSimulator}(object, simulation)

\S4method{postSimulation}{MOSimulator}(object, simulation)

\S4method{IDfromGenes}{MOSimulator}(object, geneNames, simplify = TRUE)

\S4method{IDtoGenes}{MOSimulator}(object, idNames, simplify = TRUE)

\S4method{simulateParams}{MOSimulator}(object, simulation, counts,
  profiles, group, ids)

\S4method{adjustProfiles}{MOSimulator}(object, simulation, profiles, step)

\S4method{show}{MOSimulator}(object)

\S4method{simSettings}{MOSimulator}(object)
}
\arguments{
\item{object}{Object of class \linkS4class{MOSimulator} or \linkS4class{MOSimulation}.}

\item{...}{Extra parameters for extensibility.}

\item{simulation}{Only in \linkS4class{MOSimulator} class. Initialized instance of Simulation class}

\item{geneNames}{Names of genes to look up in the association table.}

\item{simplify}{Return only the genes IDs or a table containing both types of identifiers.}

\item{idNames}{IDs to look up in the association table.}

\item{counts}{Initial count values to simulate.}

\item{profiles}{Generated simulation settings for the group being simulated.}

\item{group}{Group being simulated.}

\item{ids}{ID of every row.}

\item{step}{Two possible options "Effect" and "Groups" indicating in which point
of the settings simulation is called.}

\item{locs}{Positions of locations}

\item{chrNumber}{IDs of the chromosomes}

\item{start}{Starts positions of the regions}

\item{end}{End positions of the regions}

\item{object}{Object of class \linkS4class{MOSimulator} or \linkS4class{MOSimulation}.}

\item{object}{Instance of MOSimulator class.}

\item{simulation}{Instance of MOSimulation class.}

\item{object}{Instance of MOSimulator class.}

\item{simulation}{Instance of MOSimulation class.}

\item{object}{Instance of a MOSimulator class}

\item{object}{Instance of a simulator class.}

\item{object}{Instance of a simulator class.}

\item{simulation}{Instance of a simulation class}

\item{object}{Instance of simulator class.}

\item{simulation}{Instance of simulation class.}

\item{profiles}{Data frame containing the generated profile configuration for the
simulator.}

\item{object}{Instance of a SimulatorRegion class}

\item{object}{Instace of a SimulatorRegion class}

\item{object}{Instance of \linkS4class{MOSimulator} class.}

\item{simulation}{Initialized instance of \linkS4class{MOSimulation} class}

\item{object}{Instance of class \linkS4class{MOSimulator}.}

\item{simulation}{Instance of class \linkS4class{MOSimulation}.}

\item{object}{Instance of class \linkS4class{MOSimulator}.}

\item{geneNames}{Character vector of the genes to look for in the association table.}

\item{object}{Instance of class \linkS4class{MOSimulator}.}

\item{idNames}{Character vector of the regulator IDs to look for in the association table.}

\item{simplify}{Return only the regulator IDs or a table containing both types of identifiers.}

\item{object}{Instance of class \linkS4class{MOSimulator}.}

\item{simulation}{Instance of class \linkS4class{MOSimulation}.}

\item{counts}{Counts taken from the initial data.}

\item{profiles}{Types of profiles to simulate.}

\item{ids}{IDs associated. Needed in some simulators.}

\item{object}{Instance of class \linkS4class{MOSimulator}.}

\item{simulation}{Instance of class \linkS4class{MOSimulation}.}

\item{profiles}{Data frame containing the profile type associated to each
ID.}

\item{step}{It accepts two options \emph{Effect} and \emph{Groups} depending
on the part of the process where this method is called. Check 'Simulation.R'.}
}
\value{
A \linkS4class{MOSimulator} object containing the simulation data inside @simData
 if called on a \linkS4class{MOSimulator} class, or a \linkS4class{MOSimulation} object with all @simulators
 containing simulated data.

The simulation settings used for the general process (for MOSimulation class) or the individual omic (MOSimulator class).

An instance of MOSimulator class with modified options.

An instance of MOSimulator class with modified options.

IDs corresponding to the genes.

Genes corresponding to the IDs.

List with the following elements:
\describe{
     \item{randomCounts}{numeric vector containing random count values.}
     \item{noiseValues}{numeric vector containing noise values generated with the noise function and parameters specified.}
     \item{m}{numeric vector of lower values comparing original and random counts.}
     \item{M}{numeric vector of maximum values comparing original and random counts.}
}

The data.frame profile with the necessary modifications depending on
each child class.

GRanges object with 1 as seqname and locs as starting positions.

A character vector with the format "<chr>_<start>_<end>"

An object of class \linkS4class{MOSimulator} with the @data slot filled
with the correct initial data structure and values.

Object of class \linkS4class{MOSimulation} with the simulated data (@simData)
correctly formatted.

Depending on \emph{simplify} parameter it would be a character vector or a
data frame containing columns \emph{ID} and \emph{Gene}.

Depending on \emph{simplify} parameter it would be a character vector or a
data frame containing columns \emph{ID} and \emph{Gene}.

A list containing the following items:
\describe{
     \item{randomCounts}{numeric vector containing random count values.}
     \item{noiseValues}{numeric vector containing noise values generated with the noise function and parameters specified.}
     \item{m}{numeric vector of lower values comparing original and random counts.}
     \item{M}{numeric vector of maximum values comparing original and random counts.}
}

By default it returs the profiles data frame without modifications.
}
\description{
Generates the initial sample of data. It copies the initial data for every
group of the experimental design, and ensures to change the count values
for those genes (and associated regulators with an active effect on it)
that are marked as DE with only flat profiles in all groups.

Method to make final modifications required by the different omics. Like rounding
the final count values or renaming columns for making them more human-readable.

Returns the regulator IDs associated to a particular set of genes identifiers.

Returns the gene identifiers associated to a particular set of regulator IDs.

For internal use. Creates the values to be replaced in the formulas used
to simulate the profile values, including noise.

For internal use. Allows every omic class to adjust the profiles matrix
(simulation settings) to use, according to its needs.
}
\details{
For internal use only.
}
\section{Shared methods}{


This generic can dispatch methods for two types of classes: \describe{
\item{\linkS4class{MOSimulation}}{in which case triggers the simulation on
every simulator loaded.}
\item{\linkS4class{MOSimulator}}{performs the simulation based on already
initialized simulation settings.} }



This generic can dispatch methods from two types of classes:
\describe{
     \item{\linkS4class{MOSimulation}}{show the general simulation settings (experimental design, default depth...)}
     \item{\linkS4class{MOSimulator}}{individual omic options (depth, regulator effect...)}
}
}

\section{Simulator methods}{


This method will clone the seed data once per group to simulate and
establish differences using the simulation settings.



Method called after performing the simulation, allowing different actions
like adjusting depth, rounding values and so on, on a simulator basis.



Method for transforming genes to IDs (regions, miRNA...)



Method for transforming IDs (regions, miRNAs, ...) to genes



Method to allow individual simulator classes to change the default behaviour
of generating noise and time coefficients.



Modify the configuration profile generated when initializing simulation settings
to allow for certain constrainsts if necessary.



Creates a very basic GRanges object based on locations



Creates a valid string of a region
}

\keyword{GENERIC}
\keyword{MODIFY}
\keyword{THE}
\keyword{TODO:}
\keyword{internal}
