% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_coexpression.R
\name{make_association_dataframe}
\alias{make_association_dataframe}
\title{make_association_dataframe}
\usage{
make_association_dataframe(group, genereggroup)
}
\arguments{
\item{group}{Group from which we are generating the association dataframe}

\item{genereggroup}{list of elements to generate the association dataframe
such as clusters of each omic, indices of opposite clusters, which
genes are activated, repressed, behavior of the features etc.}
}
\value{
a dataframe with all the information the user needs about each gene
     and the order of gene and peak names to rename them in the simulated
     datasets of the group
}
\description{
This function generates a dataframe containing the information of the 
relationship between ATAC and RNA, based on the cluster groups, and then
tells the order the genes and peaks should be in the simulated dataframe
of the group
}
