% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_coexpression.R
\name{make_cluster_patterns}
\alias{make_cluster_patterns}
\title{make_cluster_patterns}
\usage{
make_cluster_patterns(numcells = 4, clusters = 8)
}
\arguments{
\item{numcells}{Number of different celltypes we are simulating}

\item{clusters}{OPTIONAL. Number of co-expression patterns the user wants
to simulate}
}
\value{
A tibble with number of columns equal to number of celltypes, rows
 according to the number of TRUE/FALSE combinations corresponding to the
 gene expression patterns along the cells
}
\description{
Function to make the tibble with cluster combinations for the gene expression
patterns along the cells
}
\examples{
patterns <- make_cluster_patterns(numcells = 4, clusters = 8)
patterns <- make_cluster_patterns(numcells = length(cell_types), 
    clusters = 8)

}
