% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{mosim}
\alias{mosim}
\title{mosim}
\usage{
mosim(omics, omicsOptions = NULL, ...)
}
\arguments{
\item{...}{Additional options to be passed to simulation object}

\item{totalGenes}{A number with the total number of genes including not expressed. Overwrited
if a genome reference is provided. Currently not used as we force to provide real data.}

\item{diffGenes}{A number with the total number of differential genes (if value > 1) or % or
total genes (if value < 1).}

\item{numberReps}{Number of replicates of the experiment.}

\item{numberGroups}{Number of samples considered on the experiment.}

\item{randomSeed}{Random seed for random number generator state.}

\item{times}{Numeric vector containing the measured times. If numberGroups < 2,
the number of times must be at least 2.}

\item{exprGenes}{Total number of expressed genes (if value > 1) or %
of the total genes (if value < 1).}

\item{noiseFunction}{Noise function to apply when simulating counts. Must accept the parameter 'n' and
return a vector of the same length. Defaults to `rnorm`}

\item{noiseParam}{Named list with the parameters to apply to the noise function.}

\item{profiles}{Named list containing the patterns with their coefficients.}

\item{profileProbs}{Numeric vector with the probabilities to assign each of the patterns. Defaults to 0.2 for each.}
}
\value{
Instance of class "Simulation" with slot "simData" containing the multiomic simulation data.
}
\description{
Performs a multiomic simulation by chaining two actions:
 1) Creating the "Simulation" class with the provided params.
 2) Calling "simulate" method on the initialized object.
}
\examples{
\dontrun{
 # Start empty simulation with default params:
 moSimulation <- mosim()

 # Retrieve simulated data, it will only contain a "SimRNAseq" key.
 moSimulation@simData


 # Simulation with every omic and some custom options:
 simulatorOptions <- list(
     'SimRNAseq'=list(
         'depth'=25
     ),
     'SimMethylseq'=list(
     ),
     'SimmiRNAseq'=list(),
     'SimDNaseseq'=list(),
     'SimChIPseq'=list()
 )

 moSimulation <- mosim(
     numberReps = 3,
     times = c(0, 2, 6, 12, 24),
     randomSeed = 1234,
     omics = simulatorOptions
 )

 # simData will contain a key for every simulator
 dataRNAseq <- moSimulation@simData$SimRNAseq

 # Methylseq is one exception, having both "beta" and "M" keys
 dataMethylSeqBeta <- moSimulation@simData$SimMethylseq$beta
}
}
