% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{omicResults}
\alias{omicResults}
\title{Retrieves the simulated data.}
\usage{
omicResults(simulation, omics = NULL, format = data.frame)
}
\arguments{
\item{simulation}{A Simulation object.}

\item{omics}{List of the omics to retrieve the simulated data.}

\item{format}{Type of object to use for returning the results}
}
\value{
A list containing an element for every omic specifiec, with
the simulation data in the format indicated.
}
\description{
Retrieves the simulated data.
}
