% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSim.R
\name{plotProfile}
\alias{plotProfile}
\title{Generate a plot of a feature's profile for one or two omics.}
\usage{
plotProfile(simulation, omics, featureIDS, drawReps = FALSE)
}
\arguments{
\item{simulation}{A Simulation object}

\item{omics}{Character vector of the omics to simulate.}

\item{featureIDS}{List containing the feature to show per omic. Must have the
omics as the list names and the features as values.}

\item{drawReps}{Logical to enable/disable the representation of the
replicates inside the plot.}
}
\value{
A ggplot2 object.
}
\description{
Generate a plot of a feature's profile for one or two omics.
}
\examples{
omic_list <- c("RNA-seq", "miRNA-seq")
rnaseq_simulation <- mosim(omics = omic_list)

plotProfile(rnaseq_simulation,
    omics = c("RNA-seq", "miRNA-seq"),
    featureIDS = list("RNA-seq"="ENSMUSG00000007682", "miRNA-seq"="mmu-miR-320-3p")
)

}
