% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsim_functions.R
\name{random_unif_interval}
\alias{random_unif_interval}
\title{random_unif_interval
Function to call the C code}
\usage{
random_unif_interval(size, max_val)
}
\arguments{
\item{size}{from sparsim}

\item{max_val}{from sparsim}
}
\description{
random_unif_interval
Function to call the C code
}
