% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_omicData}
\alias{sc_omicData}
\title{sc_omicData}
\usage{
sc_omicData(omics_types, data = NULL)
}
\arguments{
\item{omics_types}{A list of strings which can be either "scRNA-seq" or "scATAC-seq"}

\item{data}{A user input matrix with genes (peaks in case of scATAC-seq) as 
rows and cells as columns. If a user input matrix is included, cell columns 
must be sorted by cell type. Alternatively scMOSim allows you to use a default 
dataset (PBMC) by not specifying the argument.}
}
\value{
a named list with omics type as name and the count matrix as value
}
\description{
Checks if the user defined data is in the correct format, or loads the default
multiomics pbmc dataset from SeuratData package
}
\examples{
# Simulate from PBMC
omicsList <- sc_omicData(list("scRNA-seq", "scATAC-seq"))
# Simulate using data from the user
count <- omicsList[["scRNA-seq"]]
options(Seurat.object.assay.version = "v3")
Seurat_obj <- Seurat::CreateAssayObject(counts = count, assay = 'RNA')
omic_list_user <- sc_omicData(c("scRNA-seq"), c(Seurat_obj))

}
