% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{sc_param_estimation}
\alias{sc_param_estimation}
\title{sc_param_estimation}
\usage{
sc_param_estimation(
  omics,
  cellTypes,
  numberCells = NULL,
  mean = NULL,
  sd = NULL
)
}
\arguments{
\item{omics}{named list containing the omics to simulate as names, which can 
be "scRNA-seq" or "scATAC-seq".}

\item{cellTypes}{list where the i-th element of the list contains the column 
indices for i-th cell type. List must be a named list.}

\item{numberCells}{vector of numbers. The numbers correspond to the number 
of cells the user wants to simulate per each cell type. The length of the 
   vector must be the same as length of \code{cellTypes}.}

\item{mean}{vector of numbers of mean per each cell type. Must be specified 
just if \code{numberCells} is specified.}

\item{sd}{vector of numbers of standard deviation per each cell type. Must be 
specified just if \code{numberCells} is specified.}
}
\value{
a named list with simulation parameters for each omics as values.
}
\description{
Evaluate the users parameters for single cell simulation and use SPARSim
to simulate the main dataset
}
\examples{
omicsList <- sc_omicData(list("scRNA-seq"))
cell_types <- list(cellA = c(1:20), cellB = c(161:191))
estimated_params <- sc_param_estimation(omicsList, cell_types)

}
