% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{scOmicResults}
\alias{scOmicResults}
\title{scOmicResults}
\usage{
scOmicResults(sim)
}
\arguments{
\item{sim}{a simulated object from scMOSim function}
}
\value{
list of seurat objects with simulated data
}
\description{
scOmicResults
}
\examples{

cell_types <- list(cellA = c(1:20), cellB = c(161:191))
omicsList <- sc_omicData(list("scRNA-seq"))
sim <- scMOSim(omicsList, cell_types)
res <- scOmicResults(sim)
}
