% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{scOmicResults}
\alias{scOmicResults}
\title{scOmicResults}
\usage{
scOmicResults(sim)
}
\arguments{
\item{sim}{a simulated object from scMOSim function}
}
\value{
list of seurat objects with simulated data
}
\description{
scOmicResults
}
