% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{scOmicResults}
\alias{scOmicResults}
\title{scOmicResults}
\usage{
scOmicResults(sim)
}
\arguments{
\item{sim}{a simulated object from scMOSim function}
}
\value{
list of seurat objects with simulated data
}
\description{
scOmicResults
}
\examples{

cell_types <- list('Treg' = c(1:10),'cDC' = c(11:20),'CD4_TEM' = c(21:30),
'Memory_B' = c(31:40))
omicsList <- sc_omicData(list("scRNA-seq"))
sim <- scMOSim(omicsList, cell_types)
res <- scOmicResults(sim)
}
