% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_MOSim.R
\name{simulate_coexpression}
\alias{simulate_coexpression}
\title{simulate coexpression}
\usage{
simulate_coexpression()
}
\arguments{
\item{feature_ids}{}

\item{group_pattern}{}

\item{ngroups}{}

\item{sim_data}{}
}
\value{
the simulated coexpression
}
\description{
Adapted from ACORDE (https://github.com/ConesaLab/acorde) to adapt to our
data input type
}
\examples{
simulate_coexpression(sparsim_sce, 
feature_no = 3200, 
patterns, cluster_size = 400)

}
